move feed view to separate control
