+++ src/NugetPackageManager.Xaml/Views/SettingsControlView.xaml
@@ -18,7 +18,7 @@
         </Style>
     </catel:UserControl.Resources>
     <Grid>
-        <!--<Grid.Resources>
+        <Grid.Resources>
             <Style TargetType="TextBlock">
                 <Setter Property="Margin" Value="4 0 0 0"/>
             </Style>
@@ -37,7 +37,6 @@
             <RowDefinition Height="Auto"/>
         </Grid.RowDefinitions>
         <TextBlock Text="Package source" VerticalAlignment="Bottom" Margin="2 2 2 2"/>
-        <Button Command="{Binding AddFeed}">CLICK</Button>
         
         <ListView Grid.Row="1" Grid.ColumnSpan="2"
                   Grid.RowSpan="3"
@@ -70,7 +69,6 @@
                     HorizontalAlignment="Right">
             <StackPanel.Resources>
                 <Style TargetType="Button" BasedOn="{StaticResource SettingsButton}">
-                    
                 </Style>
             </StackPanel.Resources>
             <Button x:Name="AddBtn" 
@@ -83,38 +81,7 @@
                     Command="{Binding MoveDownFeed}"/>
         </StackPanel>
 
-        <TextBlock Grid.Column="2" Grid.Row="1">
-            Name:
-        </TextBlock>
-        <TextBlock Grid.Column="2" Grid.Row="2">
-            Source:
-        </TextBlock>
-
-        <TextBox Grid.Column="3" Grid.Row="1"
-                 Height="20" Width="140"
-                 Text="{Binding Name}">
-            
-        </TextBox>
-        <TextBox Grid.Column="3" Grid.Row="2"
-                 Height="20" Width="140"
-                 Text="{Binding Source}">
-            
-        </TextBox>
-
-        <Button Grid.Column="4" Grid.Row="2"
-                Height="20" Width="30"
-                Margin="4 2 2 2"
-                Command="{Binding OpenChooseLocalPathToSourceDialog}">
-            ...
-        </Button>-->
-
-        <Button Grid.Column="3" Grid.Row="3"
-                Grid.ColumnSpan="2"
-                HorizontalAlignment="Right"
-                Height="40" Width="60"
-                Margin="2 2 2 2"
-                Command="{Binding UpdateFeed}">
-            Update
-        </Button>
+        <local:FeedDetailView DataContext="{Binding SelectedFeed}"
+                              Grid.Column="3" Grid.Row="1"></local:FeedDetailView>
     </Grid>
 </catel:UserControl>
