+++ src/NugetPackageManager.Xaml/Views/SettingsWindow.xaml
@@ -6,7 +6,7 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:NugetPackageManager.Xaml.Views"
         mc:Ignorable="d"
-        Title="SettingsWindow" Height="450" Width="800">
+        Title="SettingsWindow" MinHeight="450" MinWidth="800">
     <Grid>
         <local:SettingsControlView/>
     </Grid>
